+++ common/src/main/java/it/crystalnest/fancy_entity_renderer/api/entity/player/layer/FancyCapeLayer.java
@@ -5,17 +5,17 @@ import it.crystalnest.fancy_entity_renderer.api.entity.player.model.FancyPlayerC
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.PlayerModel;
 import net.minecraft.client.model.geom.EntityModelSet;
-import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.SubmitNodeCollector;
 import net.minecraft.client.renderer.entity.RenderLayerParent;
 import net.minecraft.client.renderer.entity.layers.RenderLayer;
-import net.minecraft.client.renderer.entity.state.PlayerRenderState;
+import net.minecraft.client.renderer.entity.state.AvatarRenderState;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.resources.model.EquipmentAssetManager;
 import net.minecraft.client.resources.model.EquipmentClientInfo;
+import net.minecraft.core.ClientAsset;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.equipment.EquipmentAsset;
 import net.minecraft.world.item.equipment.Equippable;
@@ -26,16 +26,16 @@ import java.util.Optional;
 /**
  * Custom cape layer.
  */
-public class FancyCapeLayer extends RenderLayer<PlayerRenderState, PlayerModel> {
+public class FancyCapeLayer extends RenderLayer<AvatarRenderState, PlayerModel> {
   /**
    * Model for an adult player.
    */
-  private final HumanoidModel<PlayerRenderState> adultModel;
+  private final HumanoidModel<AvatarRenderState> adultModel;
 
   /**
    * Model for a baby player.
    */
-  private final HumanoidModel<PlayerRenderState> babyModel;
+  private final HumanoidModel<AvatarRenderState> babyModel;
 
   /**
    * Manager to retrieve infos about the equipment assets.
@@ -47,7 +47,7 @@ public class FancyCapeLayer extends RenderLayer<PlayerRenderState, PlayerModel> 
    * @param modelSet entity model set.
    * @param equipmentAssets {@link #equipmentAssets}.
    */
-  public FancyCapeLayer(RenderLayerParent<PlayerRenderState, PlayerModel> renderer, EntityModelSet modelSet, EquipmentAssetManager equipmentAssets) {
+  public FancyCapeLayer(RenderLayerParent<AvatarRenderState, PlayerModel> renderer, EntityModelSet modelSet, EquipmentAssetManager equipmentAssets) {
     super(renderer);
     this.adultModel = new FancyPlayerCapeModel(modelSet, false);
     this.babyModel = new FancyPlayerCapeModel(modelSet, true);
@@ -70,29 +70,19 @@ public class FancyCapeLayer extends RenderLayer<PlayerRenderState, PlayerModel> 
     return false;
   }
 
-  /**
-   * Renders this layer.
-   *
-   * @param poseStack pose stack.
-   * @param multiBufferSource buffer source.
-   * @param packedLight packed light.
-   * @param renderState render state.
-   * @param yRot rotation around the Y axis.
-   * @param xRot rotation around the X axis.
-   */
   @Override
-  public void render(@NotNull PoseStack poseStack, @NotNull MultiBufferSource multiBufferSource, int packedLight, PlayerRenderState renderState, float yRot, float xRot) {
+  public void submit(@NotNull PoseStack poseStack, @NotNull SubmitNodeCollector submitNodeCollector, int i, AvatarRenderState renderState, float v, float v1) {
     if (!renderState.isInvisible && renderState.showCape) {
-      ResourceLocation capeTexture = renderState.skin.capeTexture();
+      ClientAsset.Texture capeTexture = renderState.skin.cape();
       if (capeTexture != null && !hasLayer(renderState.chestEquipment, EquipmentClientInfo.LayerType.WINGS)) {
         poseStack.pushPose();
         if (hasLayer(renderState.chestEquipment, EquipmentClientInfo.LayerType.HUMANOID)) {
           poseStack.translate(0, -0.053125F, 0.06875F);
         }
-        HumanoidModel<PlayerRenderState> model = renderState.isBaby ? babyModel : adultModel;
-        getParentModel().copyPropertiesTo(model);
+        HumanoidModel<AvatarRenderState> model = renderState.isBaby ? babyModel : adultModel;
+//        getParentModel().copyPropertiesTo(model);
         model.setupAnim(renderState);
-        model.renderToBuffer(poseStack, multiBufferSource.getBuffer(RenderType.entitySolid(capeTexture)), packedLight, OverlayTexture.NO_OVERLAY);
+        submitNodeCollector.submitModel(model, renderState, poseStack, RenderType.entitySolid(capeTexture.texturePath()), i, OverlayTexture.NO_OVERLAY, renderState.outlineColor, null);
         poseStack.popPose();
       }
     }