+++ neoforge/src/main/java/it/crystalnest/fancy_entity_renderer/handler/package-info.java
@@ -1,4 +1,0 @@
-@ApiStatus.Internal
-package it.crystalnest.fancy_entity_renderer.handler;
-
-import org.jetbrains.annotations.ApiStatus;